Fix/remove edit team modals (#1412)
### Feature or Bugfix
<!-- please choose -->
- Bugfix

### Detail
- Originally - we were opening up a modal per TeamRow for
EnvironmentTeam and OrganizationTeam Tables whenever we wanted to edit
or remove a singular team
- For delete team in specific there was a bug when 1 team was removed
another modal remained asking to remove a separate team from the
environment (not the one user clicked on)

- This PR ensures that the Edit Team Modal or Remove Team Modal only
opens for the particular team clicked on and not for any other for both
EnvTeams and OrgTeams (prevents the above from happening)
- This design is in line with how modal showing is handled with
consumption Roles for Envs as well

### Relates
N/A

### Security
Please answer the questions below briefly where applicable, or write
`N/A`. Based on
[OWASP 10](https://owasp.org/Top10/en/).

- Does this PR introduce or modify any input fields or queries - this
includes
fetching data from storage outside the application (e.g. a database, an
S3 bucket)?
  - Is the input sanitized?
- What precautions are you taking before deserializing the data you
consume?
  - Is injection prevented by parametrizing queries?
  - Have you ensured no `eval` or similar functions are used?
- Does this PR introduce any functionality or component that requires
authorization?
- How have you ensured it respects the existing AuthN/AuthZ mechanisms?
  - Are you logging failed auth attempts?
- Are you using or adding any cryptographic features?
  - Do you use a standard proven implementations?
  - Are the used keys controlled by the customer? Where are they stored?
- Are you introducing any new policies/roles/users?
  - Have you used the least-privilege principle? How?


By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.