Redshift data sharing - Cluster encryption guardrails and information (#1447)
### Feature or Bugfix
- Feature
TO BE MERGED AFTER https://github.com/data-dot-all/dataall/pull/1424 and
after #1449

### Detail

This PR is part of the design explained in detail in #955 
- Add check that verifies the Redshift cluster imported is encrypted
before creating a redshift connection
- Store the Redshift encryption type in the redshift-connection table.
This is needed because Redshift datashares only work between namespaces
that use the same encryption type
- In the migration script and in the db model of the redshift-connection
table we add the columns encryptionType and connectionType.
ConnectionType is not needed in this particular PR, but it will be
needed in the next PR. It is added to simplify the migration scripts

### Relates
- #955 - Sharing part

### Security
Please answer the questions below briefly where applicable, or write
`N/A`. Based on
[OWASP 10](https://owasp.org/Top10/en/).

- Does this PR introduce or modify any input fields or queries - this
includes
fetching data from storage outside the application (e.g. a database, an
S3 bucket)?
  - Is the input sanitized?
- What precautions are you taking before deserializing the data you
consume?
  - Is injection prevented by parametrizing queries?
  - Have you ensured no `eval` or similar functions are used?
- Does this PR introduce any functionality or component that requires
authorization?
- How have you ensured it respects the existing AuthN/AuthZ mechanisms?
  - Are you logging failed auth attempts?
- Are you using or adding any cryptographic features?
  - Do you use a standard proven implementations?
  - Are the used keys controlled by the customer? Where are they stored?
- Are you introducing any new policies/roles/users?
  - Have you used the least-privilege principle? How?


By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.