Redshift data sharing - frontend changes in the Catalog - clean (#1458)
### Feature or Bugfix
- Feature


### Detail
Complete design in https://github.com/data-dot-all/dataall/issues/955.
This particular PR is part of the redshift dataset sharing
implementation. (It is easier to see the snapshots)
- To be able to share redshift items this PR enables the request access
button for redshift tables and datasets
- IMPORTANT: New request modal for Redshift Datasets that lists
namespaces with connections that are of type ADMIN (we cannot open a
share request for namespaces that are not "linked" to data.all). See a
screenshot of this request below:
- It adds icons in the Catalog search view for each item
- It adds the type of item right below the Catalog item
- It standardizes the icons used in the dataset list and in the Catalog
with the latests AWS public icons
- It fixes the principal type for principals that are not groups or
consumption roles
- It moves the consumption data details outside of the share view as a
child component and renders it conditionally only if the dataset type is
S3.


![Screenshot 2024-08-05 at 16 14
26](https://github.com/user-attachments/assets/35e60ada-3ce0-4fad-a1b0-c4f631624b3f)
![Screenshot 2024-08-05 at 17 45
15](https://github.com/user-attachments/assets/d7d67a9f-8387-425e-b17c-d1ae9f8745bd)

### Key decisions
- Separate request modal for redshift - we could have re-used the
request access modal, but it contains several elements that are very
particular to S3 (e.g. consumption roles, checking that the policy is
attached). So I decided to keep it separate. This way future changes to
any of the sharing methods is independent from each other.

### Relates
- #955 

### Security
Please answer the questions below briefly where applicable, or write
`N/A`. Based on
[OWASP 10](https://owasp.org/Top10/en/).

- Does this PR introduce or modify any input fields or queries - this
includes
fetching data from storage outside the application (e.g. a database, an
S3 bucket)?
  - Is the input sanitized?
- What precautions are you taking before deserializing the data you
consume?
  - Is injection prevented by parametrizing queries?
  - Have you ensured no `eval` or similar functions are used?
- Does this PR introduce any functionality or component that requires
authorization?
- How have you ensured it respects the existing AuthN/AuthZ mechanisms?
  - Are you logging failed auth attempts?
- Are you using or adding any cryptographic features?
  - Do you use a standard proven implementations?
  - Are the used keys controlled by the customer? Where are they stored?
- Are you introducing any new policies/roles/users?
  - Have you used the least-privilege principle? How?


By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.