Redshift data sharing - Redshift connection types and namespace Id (#1451)
### Feature or Bugfix
- Feature

⚠️ Needs to be merged after #1447 

### Detail
Complete design in https://github.com/data-dot-all/dataall/issues/955.
This particular PR is part of the redshift dataset sharing
implementation.

To be able to do sharing, the dataset namespace and the consumer
namespace need to be accessible by data.all. For this reason we
introduce the concept of types of connections. When opening a share
request, the datasets that can be shared are only those whose namespace
has an admin connection + the requester redshift role will be selected
from a list of namespaces with admin connections.

In addition, this PR adds the namespaceId in the RDS database for both
serverless and provisioned clusters, as it is very handy information
when working with datashares

### Relates
- #955 

### Security
Please answer the questions below briefly where applicable, or write
`N/A`. Based on
[OWASP 10](https://owasp.org/Top10/en/).

- Does this PR introduce or modify any input fields or queries - this
includes
fetching data from storage outside the application (e.g. a database, an
S3 bucket)?
  - Is the input sanitized?
- What precautions are you taking before deserializing the data you
consume?
  - Is injection prevented by parametrizing queries?
  - Have you ensured no `eval` or similar functions are used?
- Does this PR introduce any functionality or component that requires
authorization?
- How have you ensured it respects the existing AuthN/AuthZ mechanisms?
  - Are you logging failed auth attempts?
- Are you using or adding any cryptographic features?
  - Do you use a standard proven implementations?
  - Are the used keys controlled by the customer? Where are they stored?
- Are you introducing any new policies/roles/users?
  - Have you used the least-privilege principle? How?


By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.