move backend approval_tests as the last step within the backend stage (#1423)
### Feature or Bugfix
- Bugfix

### Detail
Currently if approval tests are running and a new backend deployment
happens (tests take a long time ~50 minutes so this scenario if very
likely) then the tests will run against the new deployment.
With the propose change the tests are running as part of the backend
deployment stage which is an atomic unit.


![image](https://github.com/user-attachments/assets/0068d1ac-7783-4a11-97fa-770eed21246f)


### Security
Please answer the questions below briefly where applicable, or write
`N/A`. Based on
[OWASP 10](https://owasp.org/Top10/en/).

- Does this PR introduce or modify any input fields or queries - this
includes
fetching data from storage outside the application (e.g. a database, an
S3 bucket)?
  - Is the input sanitized?
- What precautions are you taking before deserializing the data you
consume?
  - Is injection prevented by parametrizing queries?
  - Have you ensured no `eval` or similar functions are used?
- Does this PR introduce any functionality or component that requires
authorization?
- How have you ensured it respects the existing AuthN/AuthZ mechanisms?
  - Are you logging failed auth attempts?
- Are you using or adding any cryptographic features?
  - Do you use a standard proven implementations?
  - Are the used keys controlled by the customer? Where are they stored?
- Are you introducing any new policies/roles/users?
  - Have you used the least-privilege principle? How?


By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.