Redshift data sharing - Add sharing tasks to process Redshift datashares (#1467)
### Feature or Bugfix
- Feature

### Detail
Complete design in https://github.com/data-dot-all/dataall/issues/955.
This particular PR is the CORE of the redshift dataset sharing
implementation.

- Implement sharing logic in ECS task for approve, revoke and verify.
❗check the "Sharing with Redshift" section of the design with the key
decisions on the sharing workflow
- Add the necessary redshift data API calls in the redshift_data client
- Move share alarm utils to shares_base so that they can be re-used in
redshift sharing. It would be good to rename the file but it can wait.
- Includes tests for the processor functions: approve, revoke, verify

In contrast to the design in the Glue or S3 sharing mechanisms, in this
case I decided to keep it simple and use the AWS client directly from
the processor without a manager.

❗ I did not find a way to check permissions granted to redshift roles in
Redshift. For this reason in the verification task we are not checking
the last 2 steps of the share. In Redshift it is possible to check user
permissions to tables (with
[has_table_permissions](https://docs.aws.amazon.com/redshift/latest/dg/r_HAS_TABLE_PRIVILEGE.html))
and role permissions to datashares, databases and schemas woth some of
the [info tables and
views](https://docs.aws.amazon.com/redshift/latest/dg/cm_chap_system-tables.html);
but when it comes to tables there is not a table to look up or a system
function. For the moment I have not included this step, but I'll be
meeting more Redshift experts for guidance. The "good" thing is that it
is the last step for a share to succeed, so it will be a matter of users
trying and getting an "Access denied for insufficient permissions",
which can be troubleshooted

There are still a number of issues to be fixed on subsequent PRs:
- add guardrails to share creation
- polish FE (e.g. principal id, resource type)
- avoid IAM checks and dataset and IAM locks for Redshift

### Relates
- #955 

### Security
Please answer the questions below briefly where applicable, or write
`N/A`. Based on
[OWASP 10](https://owasp.org/Top10/en/).

- Does this PR introduce or modify any input fields or queries - this
includes
fetching data from storage outside the application (e.g. a database, an
S3 bucket)?
  - Is the input sanitized?
- What precautions are you taking before deserializing the data you
consume?
  - Is injection prevented by parametrizing queries?
  - Have you ensured no `eval` or similar functions are used?
- Does this PR introduce any functionality or component that requires
authorization?
- How have you ensured it respects the existing AuthN/AuthZ mechanisms?
  - Are you logging failed auth attempts?
- Are you using or adding any cryptographic features?
  - Do you use a standard proven implementations?
  - Are the used keys controlled by the customer? Where are they stored?
- Are you introducing any new policies/roles/users?
  - Have you used the least-privilege principle? How?


By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.