Allow origins conf changes (#1486)
### Feature or Bugfix
<!-- please choose -->
- Feature

### Detail
This PR is to remove wildcards from origins as reported in
https://github.com/data-dot-all/dataall/issues/1000 by using a
configuration parameter from cdk.json. This will allow the users to set
a custom origin (default is *)

### Relates
- https://github.com/data-dot-all/dataall/issues/1000

### Security
Please answer the questions below briefly where applicable, or write
`N/A`. Based on
[OWASP 10](https://owasp.org/Top10/en/).

- Does this PR introduce or modify any input fields or queries - this
includes N/A
fetching data from storage outside the application (e.g. a database, an
S3 bucket)? N/A
  - Is the input sanitized? N/A
- What precautions are you taking before deserializing the data you
consume? N/A
  - Is injection prevented by parametrizing queries? N/A
  - Have you ensured no `eval` or similar functions are used? N/A
- Does this PR introduce any functionality or component that requires
authorization? N/A
- How have you ensured it respects the existing AuthN/AuthZ mechanisms?
N/A
  - Are you logging failed auth attempts? N/A
- Are you using or adding any cryptographic features? N/A
  - Do you use a standard proven implementations? N/A
- Are the used keys controlled by the customer? Where are they stored?
N/A
- Are you introducing any new policies/roles/users? N/A
  - Have you used the least-privilege principle? How? N/A


By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.

---------

Co-authored-by: Mourya Darivemula <mouryacd@amazon.com>
Co-authored-by: Mourya Darivemula <mouryacd@80a99738cd9e.amazon.com>