Redshift data sharing - Added methods from sharing back to redshift datasets (check_on_delete, list_shared_datasets...) (#1511)
### Feature or Bugfix
- Feature

### Detail
Complete design in https://github.com/data-dot-all/dataall/issues/955.
This particular PR is focused on adding missing functionalities in
redshift_datasets that need to be implemented inside redshift_datasets.

For example, when we delete a redshift dataset we would want to first
check if there are any share requests shared for that dataset. To avoid
circular dependencies it is required to use an interface in the same way
it was implemented for S3.

In this PR:
- Add `RedshiftShareDatasetService(DatasetServiceInterface)` class and
implement required abstract methods (check_on_delete,
resolve_user_shared_datasets.....
- Use this class in redshift_Datasets module in resolvers, on dataset
deletion...
- Some of the code was very similar to the db queries implemented by S3
datasets; for this reason in this PR some of the queries are moved to
the generic ShareObjectRepository to be reused by both types of dataset

### Relates
- #955 

### Security
Please answer the questions below briefly where applicable, or write
`N/A`. Based on
[OWASP 10](https://owasp.org/Top10/en/).

- Does this PR introduce or modify any input fields or queries - this
includes
fetching data from storage outside the application (e.g. a database, an
S3 bucket)?
  - Is the input sanitized?
- What precautions are you taking before deserializing the data you
consume?
  - Is injection prevented by parametrizing queries?
  - Have you ensured no `eval` or similar functions are used?
- Does this PR introduce any functionality or component that requires
authorization?
- How have you ensured it respects the existing AuthN/AuthZ mechanisms?
  - Are you logging failed auth attempts?
- Are you using or adding any cryptographic features?
  - Do you use a standard proven implementations?
  - Are the used keys controlled by the customer? Where are they stored?
- Are you introducing any new policies/roles/users?
  - Have you used the least-privilege principle? How?


By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.