Add Glossaries integration tests (#1556)
### Feature or Bugfix
- Feature:Testing


### Detail
Implement tests for Glossaries/Catalog as part of
https://github.com/data-dot-all/dataall/issues/1220

⚠️ To test glossary associations, we need resources that use the
glossary terms (associate a term to a resource). Possible resources
include: s3_datasets, s3_tables, s3_folders, dashboards,
redshift_datasets and redshift_tables. All of which are part of other
modules that might be enabled or disabled. In this PR we assume that the
s3_datasets module is enabled! If it was not enabled, then the glossary
tests would fail!


During the implementation several enhancement ideas came up and are
collected in #1557


Tested in local with connection to real AWS deployment:

![image](https://github.com/user-attachments/assets/d759cc3d-7e44-4057-b16b-0fc94edd2290)

### Relates
- #1220 

### Security
Please answer the questions below briefly where applicable, or write
`N/A`. Based on
[OWASP 10](https://owasp.org/Top10/en/).

- Does this PR introduce or modify any input fields or queries - this
includes
fetching data from storage outside the application (e.g. a database, an
S3 bucket)?
  - Is the input sanitized?
- What precautions are you taking before deserializing the data you
consume?
  - Is injection prevented by parametrizing queries?
  - Have you ensured no `eval` or similar functions are used?
- Does this PR introduce any functionality or component that requires
authorization?
- How have you ensured it respects the existing AuthN/AuthZ mechanisms?
  - Are you logging failed auth attempts?
- Are you using or adding any cryptographic features?
  - Do you use a standard proven implementations?
  - Are the used keys controlled by the customer? Where are they stored?
- Are you introducing any new policies/roles/users?
  - Have you used the least-privilege principle? How?


By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.